修改数据类型

+++ src/main/java/com/aiqin/bms/scmp/api/purchase/domain/response/order/QueryOrderInfoRespVO.java
@@ -6,6 +6,7 @@ import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -89,19 +90,19 @@ public class QueryOrderInfoRespVO {
     private String paymentType;
 
     @ApiModelProperty("运费")
-    private Long deliverAmount;
+    private BigDecimal deliverAmount;
 
     @ApiModelProperty("商品分销价总金额")
-    private Long productTotalAmount;
+    private BigDecimal productTotalAmount;
 
     @ApiModelProperty("商品渠道价总金额")
-    private Long productChannelTotalAmount;
+    private BigDecimal productChannelTotalAmount;
 
     @ApiModelProperty("优惠额度")
-    private Long discountAmount;
+    private BigDecimal discountAmount;
 
     @ApiModelProperty("订单金额")
-    private Long orderAmount;
+    private BigDecimal orderAmount;
 
     @ApiModelProperty("商品数量")
     private Long productNum;
@@ -136,7 +137,7 @@ public class QueryOrderInfoRespVO {
     private Date operatorTime;
 
     @ApiModelProperty("活动优惠")
-    private Long activityDiscount;
+    private BigDecimal activityDiscount;
 
     @ApiModelProperty("重量")
     private Long weight;
@@ -183,10 +184,10 @@ public class QueryOrderInfoRespVO {
     private Date updateTime;
 
     @ApiModelProperty("实际渠道总价")
-    private Long actualProductChannelTotalAmount;
+    private BigDecimal actualProductChannelTotalAmount;
 
     @ApiModelProperty("实际订单金额")
-    private Long actualOrderAmount;
+    private BigDecimal actualOrderAmount;
 
     @ApiModelProperty("实际发货数量")
     private Long actualProductNum;
@@ -201,7 +202,7 @@ public class QueryOrderInfoRespVO {
     private Long actualVolume;
 
     @ApiModelProperty("实际分销总价")
-    private Long actualProductTotalAmount;
+    private BigDecimal actualProductTotalAmount;
 
     @ApiModelProperty("减免比例")
     private Integer logisticsRemissionRatio;
